BUG/MEDIUM: sink: Set the sink ref for forwarders created during ring parsing
A reference to the sink was added in every forwarder by the commit 2ae25ea24
("MINOR: sink: Add a ref to sink in the sink_forward_target structure"). But
this commit is incomplete. It is not performed for the forwarders created
during a ring parsing.

This patch must be backported to 2.6.

(cherry picked from commit 96417f392d4fdb4dca59db244a349ca0796c39b1)
Signed-off-by: Willy Tarreau <w@1wt.eu>