BUG/MEDIUM: quic: Wrong packet length check in qc_do_rm_hp()
When entering this function, we first check the packet length is not too short.
But this was done against the datagram lenght in place of the packet length.
This could lead to the header protection to be removed using data past
the end of the packet (without buffer overflow).

Use the packet length in place of the datagram length which is at <end>
address passed as parameter to this function. As the packet length
is already stored in ->len packet struct member, this <end> parameter is no
more useful.

Must be backported to 2.6.

(cherry picked from commit 99897d11d9d9c7d725bf515f2abbc94426ddfd89)
[wt: adjusted ctx in qc_lstnr_pkt_rcv]
Signed-off-by: Willy Tarreau <w@1wt.eu>