MINOR: quic: refactor datagram commit in Tx buffer
This is the final patch on quic-conn Tx refactor. Extend the function
which is used to write a datagram header to save at the same time
written buffer data. This makes sense as the two operations are used at
the same occasion when a pre-written datagram is comitted.

(cherry picked from commit 654269c769a167473d17b7bfec7c37f892975e8d)
Signed-off-by: Willy Tarreau <w@1wt.eu>