BUG/MINOR: ssl/cli: error when the ca-file is empty
"set ssl ca-file" does not return any error when a ca-file is empty or
only contains comments. This could be a problem is the file was
malformated and did not contain any PEM header.

It must be backported as far as 2.5.

(cherry picked from commit ec7eb59d206a2eb58b1d325483d196e8daaf9285)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>