BUG/MAJOR: log-forward: Fix log-forward proxies not fully initialized
Some initialisation for log forward proxies was missing such
as ssl configuration on 'log-forward's 'bind' lines.

After the loop on the proxy initialization code for proxies present
in the main proxies list, this patch force to loop again on this code
for proxies present in the log forward proxies list.

Those two lists should be merged. This will be part of a global
re-work of proxy initialization including peers proxies and resolver
proxies.

This patch was made in first attempt to fix the bug and to facilitate
the backport on older branches waiting for a cleaner re-work on proxies
initialization on the dev branch.

This patch should be backported as far as 2.3.

(cherry picked from commit 3b68b6026161c53936be34b3c837cd3febcceefc)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 42b3d6a0e04f547d35b90f405261bd53aec9badc)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>