MINOR: applet: add a function to reset the svcctx of an applet
The CLI needs to reset the svcctx between commands, and there was nothing
done to handle this. Let's add appctx_reset_svcctx() to do that, it's the
closing equivalent of appctx_reserve_svcctx().

This will have to be backported to 2.6 as it will be used by a subsequent
patch to fix a bug.

(cherry picked from commit 1cc08a33e17ca9c72204af84ecaa0c87e6d1f066)
Signed-off-by: Willy Tarreau <w@1wt.eu>