MINOR: quic: Trace fix in qc_release_frm()
This wrong trace came with this commit:
  "BUG/MINOR: quic: Possible crashes when dereferencing ->pkt quic_frame struct member"
In qc_release_frm() we mark frames as acked. Nothing to see with references
to frames.

Thank you to Willy for having caught this one.

Must be backported to 2.6 as these traces arrived with a bug fix to be backported
to 2.6.

(cherry picked from commit a846a17fdea6004e6bbbb77a1d016b4542714851)
Signed-off-by: Willy Tarreau <w@1wt.eu>