BUG/MINOR: mworker: does not create the "default" resolvers in wait mode
When doing a re-exec, the master was creating a "default" resolvers,
which could result in a warning emitted because the "default" resolvers
of the configuration file is not available anymore.

Skip the creating of the "default" resolvers in wait mode, this is not
useful in the master.

Must be backported as far as 2.6.

(cherry picked from commit 6020c4e44eb46fd21b5f85db927edb8d10e0482d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>