CLEANUP: quic: Remove a useless check in qc_lstnr_pkt_rcv()
This function parses the QUIC packet from a UDP datagram. It was originally
supposed to be run by several thread. Here we remove a section of code
where the current thread checks there is not another thread which has already
inserted the new quic_conn it is trying to insert in the connections tree.

Must be backported to 2.6 to ease the future backports to come.

(cherry picked from commit 560ddfa0035cd747c2f87d70aae2fe897dfef226)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>