BUG/MEDIUM: peers: Don't start resync on reload if local peer is not up-to-date
On a reload, if the previous resync was not finished, the freshly old worker
must not try to start a new resync. Otherwise, it will compete with the
older wokers, slowing down or blocking the resync. Only an up-to-date woker
must try to perform a local resync.

This patch must be backported as far as 2.0 (and maybe to 1.8 too).

(cherry picked from commit b372f16d352bfa20a3abbd60003e113976d16c9c)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>